+++ packages/base/src/components/Modal/Modal.styled.ts
@@ -36,7 +36,13 @@ const getModalWidthFromSize = (size: ModalContainerProps['sizeVar']) => {
   }
 };
 
-const getModalBodyTopBottomPadding = (isIncludeHeader: boolean) => {
+const getModalBodyTopBottomPadding = (isIncludeHeader: boolean, sizeVar: ModalContainerProps['sizeVar']) => {
+  if (sizeVar === 'FULL') {
+    return css`
+      padding-top: 0;
+      padding-bottom: 0;
+    `;
+  }
   if (isIncludeHeader) {
     return css`
       padding-top: 16px;
@@ -113,6 +119,7 @@ export const BodyContainer = styled.div<{
   isIncludeHeader: boolean;
   minHeight: string | number;
   maxHeight: string | number;
+  sizeVar: ModalContainerProps['sizeVar'];
 }>`
   display: flex;
   width: 100%;
@@ -121,10 +128,10 @@ export const BodyContainer = styled.div<{
   flex-direction: column;
   background: ${colorTokens.neutral0};
   box-sizing: border-box;
-  min-height: ${({ minHeight }) => minHeight};
+  min-height: ${({ minHeight }) => minHeight}px;
   max-height: ${({ maxHeight }) => maxHeight}px;
   flex: 1;
-  ${({ isIncludeHeader }) => getModalBodyTopBottomPadding(isIncludeHeader)}
+  ${({ isIncludeHeader, sizeVar }) => getModalBodyTopBottomPadding(isIncludeHeader, sizeVar)};
 `;
 
 export const ModalBodyContainerInner = styled.div`
@@ -146,6 +153,11 @@ export const ModalBodyContent = styled.div<{
   height: 100%;
   box-sizing: border-box;
   padding: 0 24px;
+  ${({ sizeVar }) =>
+    sizeVar === 'FULL' &&
+    css`
+      padding: 0;
+    `};
   background: ${colorTokens.neutral0};
 `;