+++ packages/base/src/components/index.ts
@@ -21,3 +21,4 @@ export * from './Inputs';
 export * from './ControlButtons';
 export * from './Tooltip';
 export * from './Tabs';
+export * from './Pagination';